+++ 1st-gen/packages/dialog/src/DialogWrapper.ts
@@ -23,6 +23,7 @@ import '@spectrum-web-components/button/sp-button.js';
 import '@spectrum-web-components/underlay/sp-underlay.js';
 
 // Leveraged in build systems that use aliasing to prevent multiple registrations: https://github.com/adobe/spectrum-web-components/pull/3225
+// eslint-disable-next-line import/no-extraneous-dependencies
 import '@spectrum-web-components/dialog/sp-dialog.js';
 import { Dialog } from './Dialog.js';
 import { DialogBase } from './DialogBase.js';