+++ 2nd-gen/packages/core/shared/base/Base.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2025 Adobe. All rights reserved.
+ * Copyright 2026 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -63,23 +63,25 @@ export function SpectrumMixin<T extends Constructor<ReactiveElement>>(
 ): T & Constructor<SpectrumInterface> {
     class SpectrumMixinElement extends constructor {
         /**
-         * @private
+         * @internal
          */
         public override shadowRoot!: ShadowRoot;
         private _dirParent?: HTMLElement;
 
         /**
-         * @private
+         * @internal
          */
         public override dir!: 'ltr' | 'rtl';
 
         /**
-         * @private
+         * @internal
          */
         public get isLTR(): boolean {
             return this.dir === 'ltr';
         }
-
+        /**
+         * @internal
+         */
         public hasVisibleFocusInTree(): boolean {
             const getAncestors = (root: Document = document): HTMLElement[] => {
                 // eslint-disable-next-line @spectrum-web-components/document-active-element