+++ 2nd-gen/packages/swc/.storybook/main.ts
@@ -1,6 +1,7 @@
-import { resolve, dirname } from 'path';
-import { mergeConfig } from 'vite';
+import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
+import { mergeConfig } from 'vite';
+import remarkGfm from 'remark-gfm';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -8,22 +9,40 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 const config = {
     stories: [
         {
-            directory: 'guides',
-            files: '**/*.@(md|mdx)',
-            titlePrefix: 'Guides',
-        },
-        {
             directory: '../components',
-            files: '*/stories/*.stories.ts',
+            files: '**/*.stories.ts',
             titlePrefix: 'Components',
         },
+        {
+            directory: 'learn-about-swc',
+            files: '*.mdx',
+            titlePrefix: 'Learn about SWC',
+        },
+        {
+            directory: 'guides',
+            files: '**/!(*documentation).mdx',
+            titlePrefix: 'Guides',
+        },
     ],
+    docs: {
+        defaultName: 'README',
+    },
     framework: '@storybook/web-components-vite',
     core: {
         disableTelemetry: true,
     },
     addons: [
-        '@storybook/addon-docs',
+        {
+            name: '@storybook/addon-docs',
+            options: {
+                transcludeMarkdown: true,
+                mdxPluginOptions: {
+                    mdxCompileOptions: {
+                        remarkPlugins: [remarkGfm],
+                    },
+                },
+            },
+        },
         '@storybook/addon-a11y',
         '@storybook/addon-designs',
         '@storybook/addon-vitest',